// src/helpers/types/User.ts
export type User = {
  id: string,
  givenName: string,
  picture: string,
};

export type GoogleUserInfo = {
  aud: string;
  azp: string;
  email: string;
  email_verified: boolean;
  exp: number;
  family_name: string;
  given_name: string;
  iat: number;
  iss: string;
  jti: string;
  locale: string;
  name: string;
  nbf: number;
  picture: string;
  sub: string;
};